Fix: For MSSQL apply the same workaround for the varchar size as we use for Redshift
